feat: split domain_security into passive tool + active domain_probe
domain_security bundled passive lookups (DNS/DNSSEC/CAA/email-auth via
public resolvers, RDAP via rdap.org) with three checks that touch the
target directly — AXFR zone transfer, SMTP open-relay, and the MTA-STS
policy-file fetch. Those three forced the whole tool to be classified
active, so a passive scan got no DNS/email intelligence at all.

- New apps/domain_probe (active, requires DomainAuthorization): AXFR,
  open-relay, MTA-STS fetch. source="domain_probe", check_types unchanged.
- domain_security is now passive — Passive Scan includes it (migration
  0031), so no-auth recon finally gets DNS/DNSSEC/SPF/DMARC/DKIM/RDAP.
- domain_probe joins Full Scan (0031). Registry tool count 28 -> 29.
- Reports: domain_probe scope mapping; the email-spoof CEO question
  matches both sources. domain_security findings keep their source for
  historical continuity.

Active-probe tests moved to test_domain_probe.py; domain_security's
passive tests stay. Full fast suite green (1803 passed).
