+++ apps/domain_probe/models.py
@@ -1,0 +1,5 @@
+"""Domain Probe writes to apps.core.data.findings.Finding — no models of its own.
+
+Intentionally empty: like every tool app, findings land in the unified Finding
+model. No migrations are needed because there are no models.
+"""
