+++ apps/domain_security/scanner.py
@@ -1,26 +1,25 @@
 """
-Domain security scanner for OpenEASD.
+Domain security scanner for OpenEASD — PASSIVE domain/email/RDAP intelligence.
 
-Checks:
-  - DNS: A/AAAA, NS, MX, DNSSEC, CAA, Wildcard, Zone Transfer (AXFR), Lame Delegation
-  - Email: SPF, DMARC, DKIM, MTA-STS, TLS-RPT, BIMI
+Checks (all passive — public resolvers / third-party RDAP, no packets to the
+target's own systems, so this tool needs no DomainAuthorization):
+  - DNS: A/AAAA, NS, MX, DNSSEC, CAA, Wildcard, Lame Delegation
+  - Email: SPF, DMARC, DKIM, TLS-RPT, BIMI
   - RDAP: domain expiry, transfer/delete/update locks, domain status
+
+The ACTIVE probes that touch the target directly — AXFR zone transfer, SMTP
+open-relay, and the MTA-STS policy fetch — live in ``apps.domain_probe``.
 
 All private helpers live inline in this module so that test mocks targeting
 ``apps.domain_security.scanner.*`` patch the functions that are actually called.
-The ``checks/`` subpackage contains the same logic split by concern and can be
-used independently; it is NOT imported here.
 """
 
 import datetime
 import logging
-import smtplib
 import time
 import urllib.parse
 
 import dns.resolver
-import dns.query
-import dns.zone
 import dns.message
 import dns.flags
 import dns.rcode
@@ -129,47 +128,6 @@ def _check_wildcard(session, domain) -> list:
         pass  # No wildcard — expected
     except Exception as e:
         logger.debug(f"[domain_security] Wildcard probe failed for {domain}: {e}")
-
-    return findings
-
-
-def _check_zone_transfer(session, domain, ns_records) -> list:
-    """Attempt AXFR zone transfer against each nameserver."""
-    findings = []
-
-    for ns in ns_records:
-        try:
-            ns_host = str(ns.target).rstrip(".")
-        except AttributeError:
-            ns_host = str(ns).rstrip(".")
-        try:
-            ns_ips = dns.resolver.resolve(ns_host, "A")
-            ns_ip = str(ns_ips[0])
-        except Exception:
-            continue
-
-        try:
-            zone = dns.zone.from_xfr(dns.query.xfr(ns_ip, domain, _DNS_TIMEOUT))
-            if zone:
-                record_count = sum(1 for _ in zone.nodes.keys())
-                findings.append(Finding(
-            session=session, source="domain_security", target=domain, check_type="dns",
-                    severity="critical",
-                    title=f"DNS zone transfer allowed on {ns_host}",
-                    description=(
-                        f"The nameserver {ns_host} allows unauthenticated AXFR zone transfers. "
-                        f"An attacker can enumerate all {record_count} DNS records — subdomains, "
-                        "mail servers, internal hostnames — in a single request."
-                    ),
-                    remediation=(
-                        "Restrict zone transfers to authorized secondary nameservers only. "
-                        "Configure allow-transfer ACLs on your DNS server."
-                    ),
-                    extra={"nameserver": ns_host, "record_count": record_count},
-                ))
-                break
-        except Exception:
-            pass
 
     return findings
 
@@ -353,9 +311,7 @@ def _check_dns(session, domain) -> list:
     # Wildcard DNS
     findings += _check_wildcard(session, domain)
 
-    # Zone Transfer (AXFR)
-    if ns_records:
-        findings += _check_zone_transfer(session, domain, ns_records)
+    # Zone Transfer (AXFR) is an ACTIVE probe → apps.domain_probe, not here.
 
     # Lame delegation
     if ns_records:
@@ -525,149 +481,6 @@ def _check_dkim(session, domain) -> list:
     return findings
 
 
-def _check_mta_sts(session, domain) -> list:
-    """Check MTA-STS — enforces TLS for inbound email delivery.
-
-    Two-step check per RFC 8461:
-    1. DNS TXT record at _mta-sts.domain must exist (signals policy presence)
-    2. Policy file at https://mta-sts.domain/.well-known/mta-sts.txt must be
-       reachable and have mode: enforce (mode is in the file, NOT the DNS record)
-    """
-    findings = []
-    mta_sts_records = _get_txt_record(f"_mta-sts.{domain}")
-    mta_sts_dns = next((r for r in mta_sts_records if r.startswith("v=STSv1")), None)
-
-    if not mta_sts_dns:
-        findings.append(Finding(
-            session=session, source="domain_security", target=domain, check_type="email",
-            severity="medium",
-            title="MTA-STS not configured",
-            description=(
-                f"{domain} has no MTA-STS policy. Email delivery to your mail server is not "
-                "protected against TLS downgrade attacks — a network attacker between mail "
-                "servers can force plaintext delivery and intercept email in transit."
-            ),
-            remediation=(
-                f"1. Add DNS TXT record at _mta-sts.{domain}: v=STSv1; id=<timestamp>\n"
-                f"2. Host policy file at https://mta-sts.{domain}/.well-known/mta-sts.txt\n"
-                "   Content: version: STSv1\\nmode: enforce\\nmx: <your-mx-host>\\nmax_age: 86400"
-            ),
-        ))
-        return findings
-
-    # DNS record present — fetch and validate the policy file
-    policy_url = f"https://mta-sts.{domain}/.well-known/mta-sts.txt"
-    try:
-        resp = requests.get(policy_url, timeout=_HTTP_TIMEOUT)
-        resp.raise_for_status()
-        policy_text = resp.text
-    except Exception:
-        findings.append(Finding(
-            session=session, source="domain_security", target=domain, check_type="email",
-            severity="high",
-            title="MTA-STS policy file not reachable",
-            description=(
-                f"{domain} has an MTA-STS DNS record but the policy file at {policy_url} "
-                "is not reachable. Sending mail servers cannot retrieve the policy and "
-                "will not enforce TLS — the DNS record alone provides no protection."
-            ),
-            remediation=(
-                f"Host the policy file at https://mta-sts.{domain}/.well-known/mta-sts.txt "
-                "with a valid TLS certificate. The file must be publicly accessible over HTTPS."
-            ),
-            extra={"policy_url": policy_url},
-        ))
-        return findings
-
-    # Parse mode field — this is what actually controls enforcement
-    mode = None
-    for line in policy_text.splitlines():
-        if line.strip().lower().startswith("mode:"):
-            mode = line.split(":", 1)[1].strip().lower()
-            break
-
-    if mode == "enforce":
-        return []  # correctly configured
-
-    title_map = {
-        "testing": "MTA-STS is in testing mode — TLS not enforced",
-        "none": "MTA-STS is disabled (mode: none)",
-    }
-    description_map = {
-        "testing": (
-            f"{domain} MTA-STS policy is set to mode: testing. Failures are reported "
-            "but TLS is not enforced — email can still be downgraded to plaintext."
-        ),
-        "none": (
-            f"{domain} MTA-STS policy is explicitly disabled (mode: none). "
-            "No TLS is enforced on inbound email delivery."
-        ),
-    }
-    findings.append(Finding(
-        session=session, source="domain_security", target=domain, check_type="email",
-        severity="medium",
-        title=title_map.get(mode, "MTA-STS policy mode is invalid or missing"),
-        description=description_map.get(mode, (
-            f"{domain} MTA-STS policy at {policy_url} has an unrecognised or missing "
-            f"mode field (found: {mode!r}). Sending servers will not enforce TLS."
-        )),
-        remediation="Update the MTA-STS policy file: set mode: enforce",
-        extra={"policy_url": policy_url, "mode": mode},
-    ))
-    return findings
-
-
-def _check_open_relay(session, domain) -> list:
-    """Attempt unauthenticated SMTP relay through the domain's MX server.
-
-    Connects to port 25 and sends a relay probe using two external addresses.
-    A 250 response to the RCPT TO confirms the server relays for anyone.
-    """
-    mx_records = _resolve(domain, "MX")
-    if not mx_records:
-        return []
-
-    try:
-        mx_host = str(sorted(mx_records, key=lambda r: r.preference)[0].exchange).rstrip(".")
-    except Exception:
-        return []
-
-    try:
-        with smtplib.SMTP(mx_host, 25, timeout=10) as smtp:
-            smtp.ehlo("probe.openeasd.local")
-            code, _ = smtp.mail("probe@relay-test.openeasd.local")
-            if code != 250:
-                return []
-            code, _ = smtp.rcpt("probe@relay-check.openeasd.local")
-            if code == 250:
-                return [Finding(
-                    session=session, source="domain_security", target=mx_host,
-                    check_type="open_relay", severity="critical",
-                    title="Open mail relay detected",
-                    description=(
-                        f"The mail server {mx_host} (MX for {domain}) accepted a relay "
-                        "attempt from an external address to an external address. "
-                        "Anyone on the internet can send email through this server — "
-                        "enabling spam campaigns and phishing attacks that appear to "
-                        "originate from your infrastructure, and risking IP blacklisting."
-                    ),
-                    remediation=(
-                        "Immediately restrict SMTP relay on your mail server:\n"
-                        "1. Configure your MTA to only relay for authenticated users or trusted IPs\n"
-                        "2. Postfix: set mynetworks and smtpd_relay_restrictions = permit_mynetworks, "
-                        "permit_sasl_authenticated, reject\n"
-                        "3. Exchange: disable anonymous relay in receive connectors\n"
-                        "4. Verify: telnet <mx-host> 25 → EHLO → MAIL FROM external → "
-                        "RCPT TO external → must receive 5xx rejection"
-                    ),
-                    extra={"mx_host": mx_host, "domain": domain},
-                )]
-    except Exception:
-        pass
-
-    return []
-
-
 def _check_tls_rpt(session, domain) -> list:
     """Check TLS-RPT — enables reporting of TLS failures on inbound email."""
     findings = []
@@ -735,8 +548,7 @@ def _check_email(session, domain) -> list:
     findings += _stamp_control(_check_spf(session, domain), "spf")
     findings += _stamp_control(_check_dmarc(session, domain), "dmarc")
     findings += _stamp_control(_check_dkim(session, domain), "dkim")
-    findings += _stamp_control(_check_mta_sts(session, domain), "mta_sts")
-    findings += _stamp_control(_check_open_relay(session, domain), "open_relay")
+    # MTA-STS (policy fetch) and open-relay are ACTIVE probes → apps.domain_probe.
     findings += _stamp_control(_check_tls_rpt(session, domain), "tls_rpt")
     findings += _stamp_control(_check_bimi(session, domain), "bimi")
     return findings
